# frontend_eng/api_openapi.json
{"openapi":"3.1.0","info":{"title":"Construction Inspection API","description":"API for managing construction inspections and photos","version":"1.1.0"},"paths":{"/api/projects/":{"post":{"tags":["projects"],"summary":"Create Project","description":"Create a new project","operationId":"create_project_api_projects__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Project"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["projects"],"summary":"Read Projects","description":"Get all projects, optionally filtered by owner","operationId":"read_projects_api_projects__get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}},{"name":"owner","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Owner"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Project"},"title":"Response Read Projects Api Projects  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/projects/{project_id}":{"get":{"tags":["projects"],"summary":"Read Project","description":"Get a specific project by ID with its inspections","operationId":"read_project_api_projects__project_id__get","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"integer","title":"Project Id"}},{"name":"owner","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Owner"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectWithInspections"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["projects"],"summary":"Update Project","description":"Update a project","operationId":"update_project_api_projects__project_id__put","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"integer","title":"Project Id"}},{"name":"owner","in":"header","required":true,"schema":{"type":"string","title":"Owner"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Project"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["projects"],"summary":"Delete Project","description":"Delete a project","operationId":"delete_project_api_projects__project_id__delete","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"integer","title":"Project Id"}},{"name":"owner","in":"header","required":true,"schema":{"type":"string","title":"Owner"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Project"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/projects/{project_id}/storage":{"get":{"tags":["projects"],"summary":"Get Project Storage Info","description":"獲取特定專案的靜態檔案大小資訊\n\nArgs:\n    project_id: 專案ID\n    owner: 專案擁有者 (可選)\n    db: 資料庫會話\n    \nReturns:\n    包含專案靜態檔案大小資訊的字典","operationId":"get_project_storage_info_api_projects__project_id__storage_get","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"integer","title":"Project Id"}},{"name":"owner","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Owner"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/inspections/":{"post":{"tags":["inspections"],"summary":"Create Inspection","description":"Create a new inspection","operationId":"create_inspection_api_inspections__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InspectionCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Inspection"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["inspections"],"summary":"Read Inspections","description":"Get all inspections, optionally filtered by project_id","operationId":"read_inspections_api_inspections__get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}},{"name":"project_id","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Project Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Inspection"},"title":"Response Read Inspections Api Inspections  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/inspections/{inspection_id}":{"get":{"tags":["inspections"],"summary":"Read Inspection","description":"Get a specific inspection by ID with its photos","operationId":"read_inspection_api_inspections__inspection_id__get","parameters":[{"name":"inspection_id","in":"path","required":true,"schema":{"type":"integer","title":"Inspection Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InspectionWithPhotos"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["inspections"],"summary":"Update Inspection","description":"Update an inspection result and remarks","operationId":"update_inspection_api_inspections__inspection_id__put","parameters":[{"name":"inspection_id","in":"path","required":true,"schema":{"type":"integer","title":"Inspection Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InspectionUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Inspection"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["inspections"],"summary":"Delete Inspection","description":"Delete an inspection","operationId":"delete_inspection_api_inspections__inspection_id__delete","parameters":[{"name":"inspection_id","in":"path","required":true,"schema":{"type":"integer","title":"Inspection Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Inspection"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/inspections/{inspection_id}/upload-pdf":{"post":{"tags":["inspections"],"summary":"Upload Inspection Pdf","description":"Upload a PDF for an inspection","operationId":"upload_inspection_pdf_api_inspections__inspection_id__upload_pdf_post","parameters":[{"name":"inspection_id","in":"path","required":true,"schema":{"type":"integer","title":"Inspection Id"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_inspection_pdf_api_inspections__inspection_id__upload_pdf_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Inspection"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/inspections/{inspection_id}/generate-pdf":{"post":{"tags":["inspections"],"summary":"Generate Inspection Report","description":"Generate a PDF report with inspection data and photos","operationId":"generate_inspection_report_api_inspections__inspection_id__generate_pdf_post","parameters":[{"name":"inspection_id","in":"path","required":true,"schema":{"type":"integer","title":"Inspection Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Inspection"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/photos/":{"post":{"tags":["photos"],"summary":"Create Photo","description":"Upload a new photo for an inspection","operationId":"create_photo_api_photos__post","requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_create_photo_api_photos__post"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Photo"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["photos"],"summary":"Read Photos","description":"Get all photos, optionally filtered by inspection_id","operationId":"read_photos_api_photos__get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}},{"name":"inspection_id","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Inspection Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Photo"},"title":"Response Read Photos Api Photos  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/photos/{photo_id}":{"get":{"tags":["photos"],"summary":"Read Photo","description":"Get a specific photo by ID","operationId":"read_photo_api_photos__photo_id__get","parameters":[{"name":"photo_id","in":"path","required":true,"schema":{"type":"integer","title":"Photo Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Photo"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["photos"],"summary":"Update Photo","description":"Update a photo's metadata","operationId":"update_photo_api_photos__photo_id__put","parameters":[{"name":"photo_id","in":"path","required":true,"schema":{"type":"integer","title":"Photo Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PhotoUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Photo"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["photos"],"summary":"Partial Update Photo","description":"Partially update a photo's metadata (date, caption, etc.)","operationId":"partial_update_photo_api_photos__photo_id__patch","parameters":[{"name":"photo_id","in":"path","required":true,"schema":{"type":"integer","title":"Photo Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PhotoUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Photo"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["photos"],"summary":"Delete Photo","description":"Delete a photo","operationId":"delete_photo_api_photos__photo_id__delete","parameters":[{"name":"photo_id","in":"path","required":true,"schema":{"type":"integer","title":"Photo Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Photo"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_create_photo_api_photos__post":{"properties":{"inspection_id":{"type":"integer","title":"Inspection Id"},"capture_date":{"type":"string","format":"date","title":"Capture Date"},"caption":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Caption"},"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["inspection_id","capture_date","file"],"title":"Body_create_photo_api_photos__post"},"Body_upload_inspection_pdf_api_inspections__inspection_id__upload_pdf_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_inspection_pdf_api_inspections__inspection_id__upload_pdf_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Inspection":{"properties":{"project_id":{"type":"integer","title":"Project Id"},"subproject_name":{"type":"string","title":"Subproject Name"},"inspection_form_name":{"type":"string","title":"Inspection Form Name"},"inspection_date":{"type":"string","format":"date","title":"Inspection Date"},"location":{"type":"string","title":"Location"},"timing":{"type":"string","title":"Timing"},"result":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Result"},"remark":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Remark"},"id":{"type":"integer","title":"Id"},"pdf_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Pdf Path"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["project_id","subproject_name","inspection_form_name","inspection_date","location","timing","id","created_at","updated_at"],"title":"Inspection"},"InspectionCreate":{"properties":{"project_id":{"type":"integer","title":"Project Id"},"subproject_name":{"type":"string","title":"Subproject Name"},"inspection_form_name":{"type":"string","title":"Inspection Form Name"},"inspection_date":{"type":"string","format":"date","title":"Inspection Date"},"location":{"type":"string","title":"Location"},"timing":{"type":"string","title":"Timing"},"result":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Result"},"remark":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Remark"}},"type":"object","required":["project_id","subproject_name","inspection_form_name","inspection_date","location","timing"],"title":"InspectionCreate"},"InspectionUpdate":{"properties":{"result":{"type":"string","title":"Result"},"remark":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Remark"},"pdf_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Pdf Path"}},"type":"object","required":["result"],"title":"InspectionUpdate"},"InspectionWithPhotos":{"properties":{"project_id":{"type":"integer","title":"Project Id"},"subproject_name":{"type":"string","title":"Subproject Name"},"inspection_form_name":{"type":"string","title":"Inspection Form Name"},"inspection_date":{"type":"string","format":"date","title":"Inspection Date"},"location":{"type":"string","title":"Location"},"timing":{"type":"string","title":"Timing"},"result":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Result"},"remark":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Remark"},"id":{"type":"integer","title":"Id"},"pdf_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Pdf Path"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"},"photos":{"items":{"$ref":"#/components/schemas/Photo"},"type":"array","title":"Photos","default":[]}},"type":"object","required":["project_id","subproject_name","inspection_form_name","inspection_date","location","timing","id","created_at","updated_at"],"title":"InspectionWithPhotos"},"Photo":{"properties":{"inspection_id":{"type":"integer","title":"Inspection Id"},"photo_path":{"type":"string","title":"Photo Path"},"capture_date":{"type":"string","format":"date","title":"Capture Date"},"caption":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Caption"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["inspection_id","photo_path","capture_date","id"],"title":"Photo"},"PhotoUpdate":{"properties":{"photo_path":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Photo Path"},"capture_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Capture Date"},"caption":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Caption"}},"type":"object","title":"PhotoUpdate"},"Project":{"properties":{"name":{"type":"string","title":"Name"},"location":{"type":"string","title":"Location"},"contractor":{"type":"string","title":"Contractor"},"start_date":{"type":"string","format":"date","title":"Start Date"},"end_date":{"type":"string","format":"date","title":"End Date"},"owner":{"type":"string","title":"Owner"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["name","location","contractor","start_date","end_date","owner","id"],"title":"Project"},"ProjectCreate":{"properties":{"name":{"type":"string","title":"Name"},"location":{"type":"string","title":"Location"},"contractor":{"type":"string","title":"Contractor"},"start_date":{"type":"string","format":"date","title":"Start Date"},"end_date":{"type":"string","format":"date","title":"End Date"},"owner":{"type":"string","title":"Owner"}},"type":"object","required":["name","location","contractor","start_date","end_date","owner"],"title":"ProjectCreate"},"ProjectWithInspections":{"properties":{"name":{"type":"string","title":"Name"},"location":{"type":"string","title":"Location"},"contractor":{"type":"string","title":"Contractor"},"start_date":{"type":"string","format":"date","title":"Start Date"},"end_date":{"type":"string","format":"date","title":"End Date"},"owner":{"type":"string","title":"Owner"},"id":{"type":"integer","title":"Id"},"inspections":{"items":{"$ref":"#/components/schemas/Inspection"},"type":"array","title":"Inspections","default":[]}},"type":"object","required":["name","location","contractor","start_date","end_date","owner","id"],"title":"ProjectWithInspections"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}